refine test

+++ python/paddle/fluid/initializer.py
@@ -222,7 +222,7 @@ class UniformInitializer(Initializer):
 
         assert isinstance(block, framework.Block)
         check_variable_and_dtype(var, "Out",
-                                 ["uint16", "float16", "float32", "float64"],
+                                 ["bfloat16", "float16", "float32", "float64"],
                                  "uniform_random")
 
         if self._seed == 0:
@@ -325,7 +325,7 @@ class NormalInitializer(Initializer):
         assert isinstance(block, framework.Block)
 
         check_variable_and_dtype(var, "Out",
-                                 ["uint16", "float16", "float32", "float64"],
+                                 ["bfloat16", "float16", "float32", "float64"],
                                  "guassian_random")
 
         # to be compatible of fp16 initalizers
@@ -570,7 +570,7 @@ class XavierInitializer(Initializer):
 
         assert isinstance(block, framework.Block)
         check_variable_and_dtype(var, "Out",
-                                 ["uint16", "float16", "float32", "float64"],
+                                 ["bfloat16", "float16", "float32", "float64"],
                                  "xavier_init")
 
         f_in, f_out = self._compute_fans(var)
