+++ python/paddle/fluid/layers/nn.py
@@ -10993,7 +10993,7 @@ def uniform_random_batch_size_like(input,
     check_variable_and_dtype(input, 'Input', ("float32", 'float64', "uint16"),
                              'uniform_random_batch_size_like')
     check_type(shape, 'shape', (list, tuple), 'uniform_random_batch_size_like')
-    check_dtype(dtype, 'dtype', ('float32', 'float64', "uint16"),
+    check_dtype(dtype, 'dtype', ('float32', 'float64', "bfloat16"),
                 'uniform_random_batch_size_like')
 
     helper = LayerHelper('uniform_random_batch_size_like', **locals())
@@ -15731,7 +15731,7 @@ def uniform_random(shape,
                                      float(max), 'seed', seed, 'dtype', dtype)
 
     check_type(shape, 'shape', (list, tuple, Variable), 'uniform_random/rand')
-    check_dtype(dtype, 'dtype', ('float32', 'float64', 'uint16'),
+    check_dtype(dtype, 'dtype', ('float32', 'float64', 'bfloat16'),
                 'uniform_random/rand')
 
     inputs = dict()
